feat: Adding firebase and a navigation drawer.
The login system has been added (the firebase database and rules have already been adjusted), allowing the user to log in via Google or Facebook. The layout of the main activity has been altered to contain a navigation drawer with placeholder items, which will later be replaced by the categories from the database.